requirements.txt
fastapi
uvicorn
scikit-learn
pandas
mysql-connector-python